refactor: router module
